% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{argos_prepare}
\alias{argos_prepare}
\title{argos_prepare
prepare Argos data}
\usage{
argos_prepare(dat, sf = FALSE, remove_missing_xy = TRUE)
}
\arguments{
\item{dat}{data as returned from an ARGOS table}

\item{sf}{when TRUE returns a sf. Default to FALSE.}

\item{remove_missing_xy}{when TRUE (default) removes the missing lat, lon.}
}
\description{
argos_prepare
prepare Argos data
}
\examples{
d = dbq(q = "Select * from ARGOS.2018_PESA where tagID ='52753' ")
z = argos_prepare(d)
}
