% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{cerealize}
\alias{cerealize}
\title{Serialize a complex object into a compressed character string}
\usage{
cerealize(object)
}
\arguments{
\item{object}{Any object}
}
\value{
A character string
}
\description{
Serialize a complex object into a compressed character string
}
\examples{
cerealize(mtcars)
}
