% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization.R
\name{decerealize}
\alias{decerealize}
\title{De-serialize a complex object from a compressed character string}
\usage{
decerealize(cereal)
}
\arguments{
\item{cereal}{A character string of a serialized object}
}
\value{
An R object
}
\description{
De-serialize a complex object from a compressed character string
}
\examples{
decerealize(cerealize(mtcars))
}
