% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_crw.R
\name{fitcrw}
\alias{fitcrw}
\title{fitcrw}
\usage{
fitcrw(x, frm = makefrm(x), ...)
}
\arguments{
\item{x}{a DT as obtained with argos_prepare
A thin wrapper around aniMotum::fit_ssm}
}
\description{
fitcrw
}
