% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtrack.R
\name{hashtrack}
\alias{hashtrack}
\title{Make unique hash for GPX track}
\usage{
hashtrack(track)
}
\arguments{
\item{track}{A GPX track data frame type object}
}
\value{
A SHA1 hash
}
\description{
Make unique hash for GPX track
}
\examples{
hashtrack(kaaterskill$tracks[[1]])
}
