% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kaaterskill}
\alias{kaaterskill}
\title{GPX Track File for Kaaterskill High Peak}
\format{
\subsection{\code{kaaterskill}}{

A list with a tracks entry, where tracks is a list of data frames.
}
}
\usage{
kaaterskill
}
\description{
A (simplified) GPX track file with just the track information for a hike
to Kaaterskill High Peak. Useful for configuration and testing purposes.
}
\keyword{datasets}
