% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_animap.R
\name{make_frames}
\alias{make_frames}
\title{make_frames
make_frames}
\usage{
make_frames(x, frms, baselayer, BBOX = bbox_crw(x))
}
\description{
make_frames
make_frames
}
\examples{
x = dbq(q = "Select * from ARGOS.2019_PESA where tagID = 66673") |> argos_prepare()
flagpts(x)
z = x[!(flag)]
m = fitcrw(z) |> sf::st_transform( proj4.pesa(-70) )
w = st_transform( dplyr::filter(world, continent != "Asia") ,  proj4.pesa(-70) )
ff = plan_frames(sta = min(m$date), sto = max(m$date),By = "30 mins")
out = make_frames(x = m, frms = ff, baselayer = w)
make_moviemap(out)
}
