% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_tagger.R
\name{make_point_tagger_function}
\alias{make_point_tagger_function}
\title{Make point tagger function for given list of points}
\usage{
make_point_tagger_function(points_sf, id_column = "point_id")
}
\arguments{
\item{points_sf}{An sf data frame of points}

\item{id_column}{The name of the id column}
}
\value{
A function that takes a GPX track as an argument
}
\description{
Given an sf data frame with a list of points, this will create a function
that determines which points a given track intersects. The sf data frame
must have an id_column and the required sf "geometry" column and
should be in the GPS (4326) coordinate reference system.
}
\examples{
peak_tagger <- make_point_tagger_function(peaks, "peak_id")
peak_tagger(kaaterskill$tracks[[1]])
}
