% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_crw.R
\name{makefrm}
\alias{makefrm}
\title{makefrm
make predictive frame based on time
a DT passed to aniMotum::fit_ssm time.step}
\usage{
makefrm(d, bytime = "15 mins", byid = "id")
}
\description{
makefrm
make predictive frame based on time
a DT passed to aniMotum::fit_ssm time.step
}
\examples{
x = dbq(q = "Select * from ARGOS.2019_PESA where tagID = 66673 limit 500") |>
 argos_prepare() |>
 fitcrw() |>
 st_points2lines("id")
}
