% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks}
\alias{peaks}
\title{A few peak locations}
\format{
\subsection{\code{peaks}}{

A data frame with peak_id and location.
}
}
\usage{
peaks
}
\description{
A short demo list of peaks for testing and demonstration purposes.
}
\keyword{datasets}
