% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_animap.R
\name{plan_frames}
\alias{plan_frames}
\title{plan frames
After segments were planed and defined, frames (date and file path to each frame) are defined too.}
\usage{
plan_frames(sta, sto, By = "20 mins", framedir = paste0(tempdir(), "/frames/"))
}
\description{
plan frames
After segments were planed and defined, frames (date and file path to each frame) are defined too.
}
