% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plan_segments}
\alias{plan_segments}
\title{plan segments on the map
click on points of the last leg. data is assigned to "triplegs"}
\usage{
plan_segments(d, R = 6)
}
\arguments{
\item{d}{is a sfdt output of fitcrw()}
}
\description{
plan segments on the map
click on points of the last leg. data is assigned to "triplegs"
}
