% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_animap.R
\name{prepare_segments.spatal}
\alias{prepare_segments.spatal}
\title{prepare segments.spatial
@param x a tibble from prepare_segments.temporal()}
\usage{
prepare_segments.spatal(x, polys)
}
\arguments{
\item{polys}{a list of polygons of length nrow(x)}
}
\description{
prepare segments.spatial
@param x a tibble from prepare_segments.temporal()
}
\examples{
x = prepare_segments.temporal(crsPreds = m)
p1 <- p5 <- pesa_breeding_range
p2 <- p4 <- filter(world, str_detect(continent, "America"))
p3 <- filter(world, continent == 'South America')
}
