% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_animap.R
\name{prepare_segments.temporal}
\alias{prepare_segments.temporal}
\title{prepare_segments.temporal}
\usage{
prepare_segments.temporal(x = triplegs, crsPreds)
}
\arguments{
\item{x}{default to \code{triplegs}, the object output of plan_segments()}

\item{crsPreds}{predictions from a crw model.
once marked by plan_segments(), segment start and stop are prepared}
}
\description{
prepare_segments.temporal
}
