% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{proj4.pesa}
\alias{proj4.pesa}
\title{proj4 string by project}
\usage{
proj4.pesa(
  lon_0 = {
     st_coordinates(Barrow.ll())
 }[1]
)
}
\arguments{
\item{lon0}{default to Barrow.}
}
\description{
proj4 string by project
}
