% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{tracktools-package}
\alias{tracktools}
\alias{tracktools-package}
\title{tracktools: Toools For Working With Argos And GPS Data.}
\description{
Toools for working with ARGOS and GPS data on scidb.
}
\author{
\strong{Maintainer}: Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

}
\keyword{internal}
