% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{tracktools}
\alias{tracktools}
\title{tracktools}
\description{
Toools for working with ARGOS and GPS data on scidb. Functions assume one tagID is used at a time.
}
\author{
Mihai Valcu \email{mvalcu@gwdg.de}
}
