% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{flagpts}
\alias{flagpts}
\title{flagpts
Flag outliers interactively}
\usage{
flagpts(dat, R = 10, max_speed = 500, fix_lon = TRUE)
}
\arguments{
\item{dat}{a DT returned by argos_prepare()}

\item{R}{point radius passed to leafltet::addCircleMarkers.}

\item{max_speed}{the initial maximum speed (km/h) used as a selection criterion.}

\item{fix_lon}{translate longitude so that it works with leaflet}
}
\description{
flagpts
Flag outliers interactively
}
\examples{
x <- dbq(q = "Select * from ARGOS.2018_PESA where tagID ='52753' ") |> argos_prepare()
flagpts(x)
}
